Seperated parsing logic
