+++ obc/drivers/vn100/vn100.c
@@ -5,6 +5,7 @@
 #include "obc_logging.h"
 #include "obc_sci_io.h"
 #include "vn100.h"
+#include "vn100_parsing.h"
 
 #include <FreeRTOS.h>
 #include <FreeRTOSConfig.h>
@@ -20,7 +21,6 @@
 
 #define VN100_BAUDRATE 1152000
 #define MAX_COMMAND_SIZE 256U
-#define VN100_ERR_CODE_STRING "$VNERR,"
 #define TICK_TIMEOUT portMAX_DELAY
 
 /* ---------------------------- Command Byte Sizes ----------------------------------- */
@@ -29,42 +29,6 @@
 #define ACCEL_PACKET_SIZE 12U
 #define GYRO_PACKET_SIZE 12U
 #define YMR_PACKET_SIZE 48U
-
-#define VALID_RESPONSE_STRING "$VNWRG,75"
-
-static const uint8_t PAYLOAD_OFFSET = sizeof(VALID_RESPONSE_STRING);
-static const uint16_t MAX_PAYLOAD_SIZE = sizeof(vn_ymr_packet_t);
-
-typedef enum {
-  HARD_FAULT = 1,
-  SERIAL_BUFFER_OVERFLOW = 2,
-  INVALID_CHECKSUM = 3,
-  INVALID_COMMAND = 4,
-  NOT_ENOUGH_PARAMETERS = 5,
-  TOO_MANY_PARAMETERS = 6,
-  INVALID_PARAMETER = 7,
-  INVALID_REGISTER = 8,
-  UNATHORIZED_ACCESS = 9,
-  WATCHDOG_RESET = 10,
-  OUTPUT_BUFFER_OVERFLOW = 11,
-  INSUFFICIENT_BAUD_RATE = 12,
-  ERROR_BUFFER_OVERFLOW = 255
-} VN100_error_t;
-
-typedef struct {
-  uint8_t groups;
-  uint16_t groupField;
-} VN100_header_t;
-
-typedef struct {
-  float payload[MAX_PAYLOAD_SIZE];
-} VN100_payload_t;
-
-typedef struct {
-  VN100_header_t header;
-  VN100_payload_t data;
-  uint16_t crc;
-} VN100_decoded_packet_t;
 
 /**
  * @brief Generalized request function to handle mulitple different packet types
@@ -81,10 +45,6 @@ typedef struct {
  */
 
 static obc_error_code_t serialRequestCMD(vn_cmd_t cmd, unsigned char* packet);
-static obc_error_code_t __decodePacket(vn_cmd_t cmd, unsigned char* packet, VN100_decoded_packet_t* parsedPacket);
-static obc_error_code_t parsePacket(vn_cmd_t cmd, unsigned char* packet, void* parsedPacket, VN100_error_t* error);
-static obc_error_code_t recoverErrorCodeFromPacket(unsigned char* packet, VN100_error_t* error);
-static uint16_t calculateCRC(unsigned char data[], unsigned int length);
 
 void initVN100(void) {
   sciSetBaudrate(UART_VN100_REG, VN100_BAUDRATE);
@@ -227,6 +187,7 @@ obc_error_code_t retrieveACCEL(vn_accel_packet_t* packet) {
   obc_error_code_t errCode;
   unsigned char unparsedPacket[ACCEL_PACKET_SIZE];
   RETURN_IF_ERROR_CODE(serialRequestCMD(VN_ACC, unparsedPacket));
+
   VN100_error_t error;
   RETURN_IF_ERROR_CODE(parsePacket(VN_ACC, unparsedPacket, packet, &error));
   return OBC_ERR_CODE_SUCCESS;
@@ -236,7 +197,13 @@ obc_error_code_t retrieveGYRO(vn_gyro_packet_t* packet) {
   if (packet == NULL) {
     return OBC_ERR_CODE_INVALID_ARG;
   }
-  serialRequestCMD(VN_GYR);
+
+  obc_error_code_t errCode;
+  unsigned char unparsedPacket[GYRO_PACKET_SIZE];
+  RETURN_IF_ERROR_CODE(serialRequestCMD(VN_GYR, unparsedPacket));
+
+  VN100_error_t error;
+  RETURN_IF_ERROR_CODE(parsePacket(VN_GYR, unparsedPacket, packet, &error));
   return OBC_ERR_CODE_SUCCESS;
 }
 
@@ -269,110 +236,4 @@ obc_error_code_t setASYNCOutputs(vn_cmd_t cmd) {
       break;
   }
   return OBC_ERR_CODE_SUCCESS;
-}
-
-static obc_error_code_t recoverErrorCodeFromPacket(unsigned char* packet, VN100_error_t* error) {
-  if (packet == NULL || error == NULL) return OBC_ERR_CODE_INVALID_ARG;
-
-  char errorCodePacket[] = VN100_ERR_CODE_STRING;
-  const uint8_t errorCodeIndex = sizeof(errorCodePacket) - 1;
-  const errorCode = packet[errorCodeIndex];
-
-  if (!((errorCode <= INSUFFICIENT_BAUD_RATE) || (errorCode == ERROR_BUFFER_OVERFLOW))) {
-    return OBC_ERR_CODE_VN100_PARSE_ERROR;
-  }
-  *error = errorCode;
-  return OBC_ERR_CODE_SUCCESS;
-}
-
-static obc_error_code_t __decodePacket(vn_cmd_t cmd, unsigned char* packet, VN100_decoded_packet_t* parsedPacket) {
-  if (packet == NULL || parsedPacket == NULL) return OBC_ERR_CODE_INVALID_ARG;
-
-  unsigned char* payload = packet[PAYLOAD_OFFSET];  // The main payload
-
-  VN100_decoded_packet_t decodedPacket = {0};
-  memcpy(&decodedPacket.header, payload, sizeof(decodedPacket.header));
-
-  unsigned char* data = payload[sizeof(decodedPacket.header)];
-  uint16_t packetSize = 0;
-  switch (cmd) {
-    case VN_YPR:
-      packetSize = sizeof(vn_ypr_packet_t);
-      break;
-    case VN_MAG:
-      packetSize = sizeof(vn_mag_packet_t);
-      break;
-    case VN_ACC:
-      packetSize = sizeof(vn_accel_packet_t);
-      break;
-    case VN_GYR:
-      packetSize = sizeof(vn_gyro_packet_t);
-      break;
-    case VN_YMR:
-      packetSize = sizeof(vn_ymr_packet_t);
-      break;
-    default:
-      return OBC_ERR_CODE_INVALID_ARG;
-  }
-
-  memcpy(&decodedPacket.data, data, packetSize);
-  uint16_t checksum = calculateCRC(data, packetSize);
-
-  memcpy(&decodedPacket.crc, data[packetSize], sizeof(decodedPacket.crc));
-  if (checksum != decodedPacket.crc) {
-    return OBC_ERR_CODE_VN100_PARSE_ERROR;
-  }
-
-  *parsedPacket = decodedPacket;
-  return OBC_ERR_CODE_SUCCESS;
-}
-
-static uint16_t calculateCRC(unsigned char data[], unsigned int length) {
-  unsigned int i;
-  uint16_t crc = 0;
-
-  for (i = 0; i < length; i++) {
-    crc = (unsigned char)(crc >> 8) | (crc << 8);
-    crc ^= data[i];
-    crc ^= (unsigned char)(crc & 0xff) >> 4;
-    crc ^= crc << 12;
-    crc ^= (crc & 0x00ff) << 5;
-  }
-  return crc;
-}
-
-static obc_error_code_t parsePacket(vn_cmd_t cmd, unsigned char* packet, void* parsedPacket, VN100_error_t* error) {
-  if (packet == NULL || parsedPacket == NULL) return OBC_ERR_CODE_INVALID_ARG;
-
-  /* Parsing for error */
-  char errorCodePacket[] = VN100_ERR_CODE_STRING;
-  if (!memcmp(packet, errorCodePacket, sizeof(errorCodePacket) - 1)) {
-    obc_error_code_t errCode = 0;
-    RETURN_IF_ERROR_CODE(recoverErrorCodeFromPacket(packet, error));
-    return OBC_ERR_CODE_VN100_RESPONSE_ERROR;
-  }
-
-  VN100_decoded_packet_t decodedPacket = {0};
-  obc_error_code_t errCode;
-  RETURN_IF_ERROR_CODE(__decodePacket(cmd, packet, &decodedPacket));
-
-  switch (cmd) {
-    case VN_YPR:
-      memcpy(parsedPacket, &decodedPacket.data, sizeof(vn_ypr_packet_t));
-      break;
-    case VN_MAG:
-      memcpy(parsedPacket, &decodedPacket.data, sizeof(vn_mag_packet_t));
-      break;
-    case VN_ACC:
-      memcpy(parsedPacket, &decodedPacket.data, sizeof(vn_accel_packet_t));
-      break;
-    case VN_GYR:
-      memcpy(parsedPacket, &decodedPacket.data, sizeof(vn_gyro_packet_t));
-      break;
-    case VN_YMR:
-      memcpy(parsedPacket, &decodedPacket.data, sizeof(vn_ymr_packet_t));
-      break;
-    default:
-      return OBC_ERR_CODE_INVALID_ARG;
-  }
 }
