+++ test/test_obc/unit/test_vn100_packet_parsing.cpp
@@ -1,0 +1,6 @@
+#include "vn100_parsing.h"
+#include "obc_errors.h"
+
+#include <gtest/gtest.h>
+
+TEST(VN100_PARSING_SUITE, EXPECT_SERIAL_ERROR_CODE) {}
